+++ src/internal/api/auth/infrastructure/handlers/get_user_handler_test.go
@@ -31,7 +31,7 @@ func TestGetUserHandler_Returns_An_Error_If_The_Query_Fails(t *testing.T) {
 	mockedRepo := repository.MockedUsersRepository{}
 	h := handler.Handler{UsersRepository: &mockedRepo}
 
-	mockedRepo.On("FindUser", request().Context(), domain.UserEntity{ID: int32(1)}).Return(nil, fmt.Errorf("some error")).Once()
+	mockedRepo.On("FindUser", request().Context(), domain.UserEntity{ID: 1}).Return(nil, fmt.Errorf("some error")).Once()
 
 	result := GetUserHandler(httptest.NewRecorder(), request(), h)
 
@@ -53,7 +53,7 @@ func TestGetUserHandler_Returns_The_User(t *testing.T) {
 
 	nvo, _ := domain.NewUserNameValueObject("user1")
 	user := domain.UserEntity{ID: 2, Name: nvo, IsAdmin: true}
-	mockedRepo.On("FindUser", request().Context(), domain.UserEntity{ID: int32(1)}).Return(&user, nil).Once()
+	mockedRepo.On("FindUser", request().Context(), domain.UserEntity{ID: 1}).Return(&user, nil).Once()
 
 	result := GetUserHandler(httptest.NewRecorder(), request(), h)
